A11Y: add `aria-hidden="true"` wrapper around zero width space character in buttons (#31544)
The zero width space added here as a layout fix seems to get noticed by
screenreaders (tested in NVDA) and gets "read" as text.

This means that instead of falling back to the button title, which is
the normal behavior for textless buttons, the screenreader reads the
blank space. This results in buttons like the post controls being read
simply as "button" with no other description.

Wrapping the space in `aria-hidden` corrects this, and results in the
button title being read properly.