A11Y: retain focus on highlighted post to avoid scroll jump (#31547)
Here we correctly focus a mid-stream post for screenreaders so they know
where to start, but once focus is removed NVDA becomes unmoored and will
jump back up the post stream to a previous post if someone tries to
navigate with the arrow keys.

Retaining the focus prevents this and shouldn't have any major
side-effects.

Tested using Chrome and NVDA