DEV: correct class on trigger of notifications-tracking
If we have a custom content for a `DMenu` trigger the `DButton` has no way to know there's text in this custom content so it will output the "no-text" class on the `DButton`. The correct way to fix this is to use our own `DButton` and the `@triggerComponent` so we can fully replace the behavior and set the correct classes.