DEV: aliases should be considered as existing
Before this commit doing: `Emoji.exists?(some_alias)` would return false. The only important thing in emoji is to never remove an emoji code which has been used by users but changing the name of an emoji and keeping old name as an alias should not break the application in any way, this commit should ensure this is true.