DEV: set field type in control wrapper
Calling `<field.Calendar>` will ultimately call `componentFor` which will mutate the type of the field. Problem is the way we use a hash to define all the available controls, would work correctly on first render, but because of the mutation in componentFor if something would trigger a re-render it would have to evaluate all the componentFor calls, and the last one would "win" (currently calendar), which would mean that a field rendered as custom would for example become calendar.

The easiest repro was:

```gjs
<Form as  |form|>
  <form.field as |field|>
    {{log field}}
    <field.Custom>
      Hello world
    </field.Custom>
  </form.field>
</Form>
```

The fix for now is to mutate the value at the point where the component is actually rendered, in control-wrapper.

A possibly more correct longer term fix could be to change the API, to set the type on the field when defined in the template, and render a generic component which would use this type to know what to render:

```
<form.field @type="calendar" as |field|>
  <field.Control @someCalendarOption="bar" />
</form.field>
```