DEV: Remove installing minio binaries in tests workflow
In the `discourse/discourse_test:release` base image, the minio binaries
are already being installed and updated on a daily basis. There is also no
real need for us to have to run the latest minio binaries once it is
released. As such, we can drop the step in the `tests` workflow.