FIX: Show total members count when deleting a group (#31699)
When deleting a group, we show a confirmation dialog that includes the
group members count that will lose access to the group alongside other
information. However, there's currently an issue where the members count
shown in the dialog is the count of members that have been loaded in the
Members tab of the group. If you filter the members list and then click
delete, the count you currently see is the number of members that match
the filter, not the total members count.

This happens because the dialog uses the length of the members list
(`members.length`) instead of the `user_count` attribute of the group.

Internal topic: t/148832.