DEV: Add protection for migrations that creates index concurrently
This commit updates `Migration::SafeMigrate` to protect against unsafe
ways of adding a Postgres index concurrently.

Per postgres documentation:

  If a problem arises while scanning the table, such as a deadlock or a uniqueness violation in a unique index,
  the CREATE INDEX command will fail but leave behind an "invalid" index. This index will be ignored for querying
  purposes because it might be incomplete; however it will still consume update overhead. The recommended recovery
  method in such cases is to drop the index and try again to perform CREATE INDEX CONCURRENTLY .

Therefore, the simplest way for us to ensure that migrations that create
indexes concurrently are idempotent is to follow postgres'
recommendation of dropping the index first before trying to create the
index concurrently.