FEATURE: enable full page login by default
This switches the signup/login UI to the full page experience by default.

It also marks the setting as hidden. THe plan is to keep the setting for
only a short period of time to allow for quickly reverting if there are
any major issues. After that, the setting will be removed.