DEV: removes code block enter edge case
This was causing some confusion and now that we have the chat preference to send with enter or metakey+enter it's even more confusing.

If your setting is send on enter, you need to use shift+enter for a new line, even in codeblocks. If your setting is meta+enter you need to use enter in the codeblock for a new line, meta+enter will send as expected.