WIP: Fix inline spoilers when using toolbar Blur Spoiler
When we do `applySurround`, if we have an active ProseMirror
selection via this.view.state.selection, then we should not use
the "example text" when attempting to add [spoiler] bbcode.

In addition to this, when we call `convertFromMarkdown` we must
add an additional bit of text at the beginning otherwise the
parser will always parse e.g. `[spoiler]test[/spoiler]` as
a spoiler block bbcode rather than inline bbcode