PERF: Optimize topic-tracking-state subcategory calculation
When opening the sidebar on a site with thousands of categories, `getSubCategoryIds` is called repeatedly and the nested loops burn a lot of CPU time.

The `Category.descendants` logic is similar in efficiency, but crucially it is cached so it only needs to be calculated once per category. So this change should make a big difference to sidebar performance on sites with lots of categories.