PERF: Optimize `Category.subcategories`
At the moment, every call to `Category.subcategories` causes an iteration through every single category. For sites with thousands of categories, this is incredibly expensive, and makes things like the category dropdown & sidebar rendering very slow.

To improve things, we can create a map of parent_category => [child categories] in a single iteration, cache it, and then use that when finding subcategories.

The `@computed` decorator ensures that this cache will be recalculated whenever the list of categories changes (e.g. when sideloading new categories in the experimental lazy-loaded-categories mode)