DEV: add value transformer to allow bulk select in nav controls (#31800)
It's somewhat common for a theme to remove our topic list header: 

![image](https://github.com/user-attachments/assets/28fbe764-0944-4d96-b84e-00fd68fda385)

One issue with this is that it eliminates the bulk select button. 

Fortunately, we already have a solution for this on mobile... show it in
the navigation-controls:


![image](https://github.com/user-attachments/assets/17d0b20a-2948-4f3e-8615-7f8577adba3d)


This PR adds the `showBulkSelectInNavControls` value transformer — so we
can optionally allow this on desktop as well:


![image](https://github.com/user-attachments/assets/883975ec-95ed-44b4-85cc-6c9aa53142cd)

This provides an easy way to keep the bulk select controls without the
topic list header.

```js
  api.registerValueTransformer("bulk-select-in-nav-controls", () => {
    return true;
  });
 ```