UX: Improve loading state when changing admin user list (#31856)
Previously we were showing a loading spinner, but the old user list
persisted so the spinner was often off the bottom of the screen. This
commit updates the `users` list to be a getter, so it's always perfectly
in-sync with the `_results` set. That means no users will be shown while
a new list is being loaded, so the spinner is more visible.


https://meta.discourse.org/t/load-spinner-missing-from-dynamic-pages/357525/4