DEV: Add polyfill for CSS relative color syntax
This polyfill only works with static color values, which means it will NOT work on things like `olclh(from var(--primary) ...`.

Therefore, we must only use it in color_definition stylesheets, and use scss variables as the source color. Like `okclh(from #{$primary}`.

Once we drop iOS 15, we'll be able to drop this polyfill, and use the syntax anywhere with `var()`