FIX: Add value transformers for welcome banner and search
We want to give some themes more control over
the default experience in core. To this end, this
commit adds value transformers for these
site settings:

* enable_welcome_banner
* search_experience

This way, themes can give a different value to
what has been set in core.

This can be potentially confusing, a follow up PR
may refactor this a bit so there is a more direct
correlation between the setting and the transformer
so we could show a warning in the site setting page
for example.