DEV: Move dialog DOM handling from service to component
Accessing a component-rendered DOM element from a service is not ideal, and can lead to timing issues. We've attempted to work around this using `afterRender`, but it's still not ideal.

This commit moves the `A11YDialog` setup into a modifier. Now the service only stores state, and the component renders things based on that state. This is much more in line with Ember norms, and should be much more robust.

Test adjustments are because the dialog element is now only rendered on-demand, rather than being present in the DOM all the time. This brings dialog inline with our Modal system.