DEV: adds declarative options to object/container (#31934)
We often need to be able to change the format of a container and to
change the flex-direction, this is now possible through two properties:

`<form.Container @format="large">`
`<form.Container @direction="column">`

On top of this `Object` has now a similar behavior to `Collection`. It
will be displayed as flex, add gap between children and accepts
`...attributes`.